requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-huggingface==0.0.1
langchain-ollama==0.1.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
